Automating AR 1-* TITLE.html
In this update, I have begun rewriting the TITLE.html files in individual AR folders.  The real key is that I have written a spreadsheet model to fill in the variable code.  I have limited this initial push to the AR 1* Series (about 10 files) to make sure that it works.

Additional minor changes
- With each CURRENT folder, I am deleting the FIG folder...that was a hold over from the time when I was planning on digesting an entire regulation and recreating it as HTML.
-Also, I renamed the file REDIR.html file in 1-201 (that was also an earlier attempt at managing connections between documents.)